requirements.txt
gitpython>=3.1.0
langchain>=0.0.208
openai>=1.0.0
python-dotenv>=1.0.0
click>=8.0.0
PyYAML>=6.0.0
